feat(UNT-T8286): implement card swipe animation component (with hooks)
