+++ example/src/constants/index.ts
@@ -1,0 +1,3 @@
+import {imageData} from './DummyData';
+
+export {imageData};
