+++ src/components/CardSwipeable/hooks/index.ts
@@ -1,0 +1,3 @@
+import useAnimatedCard from './useAnimatedCard';
+
+export { useAnimatedCard };
