+++ src/components/CardSwipeable/index.ts
@@ -1,8 +1,15 @@
-import CardSwipeable from './CardSwipeable';
+import SimpleCard from './SimpleCard';
+import SwipeableCard from './SwipeableCard';
 
-export { CardSwipeable };
+const CardSwipeable = Object.assign(SimpleCard, {
+  SimpleCard: SimpleCard,
+  SwipeableCard: SwipeableCard,
+});
+
 export type {
   SimpleCardProps,
   SwipeableCardProps,
   SwipeCallBackProps,
+  SwipeCallEndProps,
 } from './types';
+export { CardSwipeable };
