--- src/components/CounterButton/__tests__/CounterButton.test.tsx
@@ -1,13 +1,0 @@
-//@ts-nocheck
-import React from 'react';
-import { render } from '@testing-library/react-native';
-import { CounterButton } from '../CounterButton';
-
-jest.useFakeTimers();
-
-describe('CounterButton component', () => {
-  it('Match Snapshot', () => {
-    const { toJSON } = render(<CounterButton />);
-    expect(toJSON()).toMatchSnapshot();
-  });
-});
