--- src/components/CounterButton/index.ts
@@ -1,4 +1,0 @@
-import CounterButton from './CounterButton';
-
-export { CounterButton };
-export type { CounterButtonProps } from './types';
