--- src/components/CounterButton/styles.ts
@@ -1,11 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
-
-export default styles;
