--- src/components/CounterButton/types.ts
@@ -1,3 +1,0 @@
-export interface CounterButtonProps {
-  label: string;
-}
