+++ src/index.ts
@@ -1,4 +1,3 @@
-export * from './components/CounterButton';
 export * from './components/GradientProgress';
 export * from './components/CardSwipeable';
 export * from './theme';
